rename files according to linting rules
