+++ meta_configurator/.ls-lint.yml
@@ -1,7 +1,9 @@
 ls:
   src:
-    .dir: kebab-case
-    .js: kebab-case
+    .dir: kebab-case | regex:__tests__
+    .js: camelCase
+    .ts: camelCase | kebab-case
+    .test.ts: camelCase
     .css: regex:([.a-z]*)([-.][a-z]+)*
     .vue: PascalCase
