+++ meta_configurator/src/components/code-editor/codeEditorWrapperAce.ts
@@ -1,4 +1,4 @@
-import type {CodeEditorWrapper} from '@/components/code-editor/CodeEditorWrapper';
+import type {CodeEditorWrapper} from '@/components/code-editor/codeEditorWrapper';
 import type {Editor} from 'brace';
 
 /**
