+++ meta_configurator/src/components/gui-editor/PropertyData.vue
@@ -3,7 +3,7 @@ Wrapper component that contains the actual component that allows the user to edi
 The component is chosen based on the type of the property, see resolveCorrespondingComponent.
 -->
 <script setup lang="ts">
-import type {ConfigTreeNodeData} from '@/model/ConfigDataTreeNode';
+import type {ConfigTreeNodeData} from '@/model/configDataTreeNode';
 import type {Path} from '@/model/path';
 import {resolveCorrespondingComponent} from '@/components/gui-editor/resolveCorrespondingComponent';
 import {pathToString} from '@/utility/pathUtils';
