+++ meta_configurator/src/components/gui-editor/properties/AnyOfSelectionProperty.vue
@@ -5,11 +5,11 @@ Component for selecting one or more sub-schemas of an anyOf property.
 import type {WritableComputedRef} from 'vue';
 import {computed} from 'vue';
 import MultiSelect from 'primevue/multiselect';
-import type {JsonSchema} from '@/schema/JsonSchema';
+import type {JsonSchema} from '@/schema/jsonSchema';
 import {useSessionStore} from '@/store/sessionStore';
 import type {Path, PathElement} from '@/model/path';
 import {pathToString} from '@/utility/pathUtils';
-import {OneOfAnyOfSelectionOption, schemaOptionToString} from '@/model/OneOfAnyOfSelectionOption';
+import {OneOfAnyOfSelectionOption, schemaOptionToString} from '@/model/oneOfAnyOfSelectionOption';
 
 const props = defineProps<{
   propertyName: PathElement;
