+++ meta_configurator/src/components/gui-editor/properties/__tests__/enumProperty.test.ts
