+++ meta_configurator/src/example-schemas/autonomousVehicleSchema.ts
@@ -1,4 +1,4 @@
-import type {TopLevelSchema} from '@/model/JsonSchemaType';
+import type {TopLevelSchema} from '@/model/jsonSchemaType';
 
 export const AUTONOMOUS_VEHICLE_SCHEMA: TopLevelSchema = {
   $schema: 'https://json-schema.org/draft/2020-12/schema',
