+++ meta_configurator/src/example-schemas/schemaCollection.ts
@@ -1,8 +1,8 @@
-import {DEFAULT_SCHEMA} from '@/example-schemas/DefaultSchema';
-import {STRENDA_SCHEMA} from '@/example-schemas/StrendaSchema';
-import {ENZYMEML_SCHEMA} from '@/example-schemas/EnzymemlSchema';
-import {AUTONOMOUS_VEHICLE_SCHEMA} from '@/example-schemas/AutonomousVehicleSchema';
-import type {SchemaOption} from '@/model/SchemaOption';
+import {DEFAULT_SCHEMA} from '@/example-schemas/defaultSchema';
+import {STRENDA_SCHEMA} from '@/example-schemas/strendaSchema';
+import {ENZYMEML_SCHEMA} from '@/example-schemas/enzymemlSchema';
+import {AUTONOMOUS_VEHICLE_SCHEMA} from '@/example-schemas/autonomousVehicleSchema';
+import type {SchemaOption} from '@/model/schemaOption';
 
 /**
  * Collection of all available example schemas.
