+++ meta_configurator/src/example-schemas/settingsSchema.ts
