+++ meta_configurator/src/model/jsonSchemaType.ts
