+++ meta_configurator/src/model/schemaOption.ts
@@ -1,4 +1,4 @@
-import type {TopLevelSchema} from '@/model/JsonSchemaType';
+import type {TopLevelSchema} from '@/model/jsonSchemaType';
 
 export interface SchemaOption {
   label: string;
