+++ meta_configurator/src/model/settingsTypes.ts
