+++ meta_configurator/src/schema/jsonSchema.ts
@@ -1,5 +1,5 @@
-import type {JsonSchemaType, SchemaPropertyType} from '@/model/JsonSchemaType';
-import {nonBooleanSchema, schemaArray, schemaFromObject, schemaRecord} from '@/schema/SchemaUtils';
+import type {JsonSchemaType, SchemaPropertyType} from '@/model/jsonSchemaType';
+import {nonBooleanSchema, schemaArray, schemaFromObject, schemaRecord} from '@/schema/schemaUtils';
 import type {Path, PathElement} from '@/model/path';
 import {preprocessSchema} from '@/schema/schemaPreprocessor';
 import _ from 'lodash';
