--- meta_configurator/src/test/dummy.test.ts
@@ -1,7 +1,0 @@
-import {describe, expect, test} from 'vitest';
-
-describe('dummy', () => {
-  test('dummy', () => {
-    expect(1).toBe(1);
-  });
-});
